Part4.21

+++ part4/bloglist/controllers/blogs.js
@@ -36,8 +36,22 @@ blogRouter.post('/', async (request, response, next) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+
+    const blog = await Blog.findById(request.params.id)
+
+    if ( blog.user.toString() === user.id.toString() ){
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    }else{
+        return response.status(401).json({ error: 'token invalid' })
+    }
+
 })
 
 blogRouter.put('/:id', async (request, response, next) => {
